step1PRCollection
