+++ CodeStarterCamp_Week3/main.swift
@@ -8,39 +8,7 @@
 
 import Foundation
 
-struct Person {
-  var money: Int
-  init(money: Int) {
-    self.money = money
-  }
-    func buy(){
-      print("물건을 구매합니다.")
-  }
-}
-
-
-struct CoffeeShop {
-  var totalSales = 0
-  var barista: Person?
-  let menu: [String: Int] = ["americano": 100, "cappuchino": 120]
-  var pickUpTable = "pickUpTable"
-  func takeOrder(){
-    print("주문을 받습니다.")
-  }
-  func makeCoffee(){
-    print("커피를 만듭니다.")
-  }
-}
-
-
-enum Coffee {
-  case milk
-  case iceamricano
-  case cappuccino
-}
-
-
-var yagombucks: CoffeeShop = CoffeeShop()
+var yagombucks: CoffeeShop = CoffeeShop(totalSales: 0, pickupTable: "pickupTable")
 let misterKim: Person = Person(money: 0)
 
 yagombucks.barista = misterKim
