+++ CodeStarterCamp_Week3/person.swift
@@ -1,0 +1,18 @@
+//
+//  person.swift
+//  CodeStarterCamp_Week3
+//
+//  Created by song on 2021/12/23.
+//
+
+import Foundation
+
+class Person {
+  var money: Int
+  init(money: Int) {
+    self.money = money
+  }
+    func buy(){
+      print("물건을 구매합니다.")
+  }
+}
